requirements.txt
torch>=2.0.0
torchvision>=0.15.0
albumentations
timm==0.9.12
mediapipe==0.10.14
opencv-python>=4.7.0
gradio>=4.0.0
numpy>=1.23.0
pillow>=9.5.0
